SNAP-1351 (#9)
Changes proposed in this pull request

When starting a job, a property is set which decides whether to interrupt task threads if the job/tasks are killed. This property is set as true by job server. During a failure, when tasks are killed by the driver, it sends executors a message to interrupt the task threads. This causes issues with snappy because snappy may be writing to an oplog and it generates a DiskAccessException. This DAE ends up closing the underlying regions.

Snappy threads cannot be interrupted and hence making this flag as false.
Patch testing

Tried out the scenario mentioned in the bug. Works fine. Precheckin.
Other PRs

SnappyDataInc/snappydata#581